refactor: change blobs to a function
