+++ src/features/home/pages/home-page.tsx
@@ -1,6 +1,6 @@
 import { Background } from '@/shared/components/background'
 import { Layout } from '@/shared/components/layout'
-import { motion, Transition } from 'framer-motion'
+import { motion } from 'framer-motion'
 import { Button } from '@/shared/components/button'
 import { Accordion, AccordionItem } from '@/shared/components/accordion'
 import { DotLottieReact } from '@lottiefiles/dotlottie-react'
@@ -19,6 +19,7 @@ import {
   CardTitle
 } from '@/shared/components/card'
 import { ArrowDown, ArrowRight, Check } from 'lucide-react'
+import { BackgroundBlobs } from '@/shared/components/background-blobs'
 
 const plans = [
   {
@@ -96,56 +97,6 @@ const process = [
   }
 ]
 
-function BackgroundBlobs() {
-  const blobTransition = {
-    repeat: Infinity,
-    repeatType: 'mirror',
-    duration: 30,
-    ease: 'easeInOut'
-  } as const satisfies Transition
-
-  return (
-    <>
-      {/* Light mode blobs use CSS variables for consistent visibility */}
-      <motion.div
-        initial={{ opacity: 0, scale: 0.5 }}
-        animate={{
-          opacity: [0, 0.4, 0.2, 0.4, 0],
-          x: [0, -100, 0],
-          y: [0, 50, 0]
-        }}
-        transition={blobTransition}
-        className="fixed inset-0 -z-10 m-auto h-[80vh] w-[80vw] rounded-full blur-[200px]"
-        style={{ backgroundColor: 'var(--primary)', opacity: 0.3 }}
-      />
-
-      <motion.div
-        initial={{ opacity: 0, scale: 0.6 }}
-        animate={{
-          opacity: [0, 0.3, 0.1, 0.3, 0],
-          x: [50, -50, 50],
-          y: [20, -20, 20]
-        }}
-        transition={{ ...blobTransition, duration: 35 }}
-        className="fixed inset-0 -z-20 m-auto h-[60vh] w-[60vw] rounded-full blur-[180px]"
-        style={{ backgroundColor: 'var(--secondary)', opacity: 0.2 }}
-      />
-
-      <motion.div
-        initial={{ opacity: 0, scale: 0.4 }}
-        animate={{
-          opacity: [0, 0.2, 0.05, 0.2, 0],
-          x: [-50, 50, -50],
-          y: [50, -50, 50]
-        }}
-        transition={{ ...blobTransition, duration: 40 }}
-        className="fixed inset-0 -z-30 m-auto h-[50vh] w-[50vw] rounded-full blur-[220px]"
-        style={{ backgroundColor: 'var(--accent)', opacity: 0.25 }}
-      />
-    </>
-  )
-}
-
 function LogoCarousel() {
   const logos = [
     'https://store-images.s-microsoft.com/image/apps.22898.1ee6619e-8cd6-4c91-a459-3abe6a6cf7b3.84da202a-9b47-4a76-be2c-79f143c64050.21add500-ed5b-42ee-b87d-ee415ae071fc',
@@ -202,7 +153,7 @@ const words = title.split(' ')
 
 export function HomePage() {
   return (
-    <Background className="to-background/90 from-background relative isolate flex flex-col items-center justify-center overflow-hidden bg-gradient-to-b">
+    <Background className="relative isolate overflow-hidden">
       <BackgroundBlobs />
       <Layout className="space-y-32 md:space-y-0">
         <motion.section
